+++ Sources/Sentry/SentryScope.m
@@ -1,5 +1,6 @@
 #import "NSMutableDictionary+Sentry.h"
 #import "SentryAttachment+Private.h"
+#import "SentryBreadcrumb+Private.h"
 #import "SentryBreadcrumb.h"
 #import "SentryDefines.h"
 #import "SentryEvent+Private.h"
@@ -9,13 +10,15 @@
 #import "SentryLogC.h"
 #import "SentryScope+Private.h"
 #import "SentryScope+PrivateSwift.h"
-#import "SentrySpanInternal+Private.h"
+#import "SentrySpanInternal.h"
 #import "SentrySwift.h"
 #import "SentryTracer.h"
 #import "SentryTransactionContext.h"
 #import "SentryUser.h"
 
 NS_ASSUME_NONNULL_BEGIN
+
+static NSString *const kSentryScopeSpanStatusSerializationKey = @"status";
 
 @interface SentryScope ()
 
@@ -36,6 +39,7 @@ NS_ASSUME_NONNULL_BEGIN
     NSObject *_spanLock;
     NSObject *_observersLock;
     NSObject *_propagationContextLock;
+    SentryFeatureFlagBufferWrapper *_featureFlagBuffer;
 }
 
 @synthesize span = _span;
@@ -55,6 +59,7 @@ NS_ASSUME_NONNULL_BEGIN
         self.attachmentArray = [[NSMutableArray alloc] init];
         self.fingerprintArray = [[NSMutableArray alloc] init];
         self.attributesDictionary = [[NSMutableDictionary alloc] init];
+        _featureFlagBuffer = [SentryFeatureFlagBufferWrapper scopeBuffer];
         _spanLock = [[NSObject alloc] init];
         _observersLock = [[NSObject alloc] init];
         _propagationContextLock = [[NSObject alloc] init];
@@ -69,9 +74,14 @@ NS_ASSUME_NONNULL_BEGIN
     return [self initWithMaxBreadcrumbs:defaultMaxBreadcrumbs];
 }
 
+- (SentryFeatureFlagBufferWrapper *)featureFlagBuffer
+{
+    return _featureFlagBuffer;
+}
+
 - (instancetype)initWithScope:(SentryScope *)scope
 {
-    if (self = [self init]) {
+    if (self = [self initWithMaxBreadcrumbs:scope.maxBreadcrumbs]) {
         [_extraDictionary addEntriesFromDictionary:[scope extras]];
         [_tagDictionary addEntriesFromDictionary:[scope tags]];
         [_contextDictionary addEntriesFromDictionary:[scope context]];
@@ -82,6 +92,7 @@ NS_ASSUME_NONNULL_BEGIN
         [_fingerprintArray addObjectsFromArray:[scope fingerprints]];
         [_attachmentArray addObjectsFromArray:[scope attachments]];
         [_attributesDictionary addEntriesFromDictionary:[scope attributes]];
+        _featureFlagBuffer = [scope.featureFlagBuffer copyBuffer];
 
         self.propagationContext = scope.propagationContext;
         self.maxBreadcrumbs = scope.maxBreadcrumbs;
@@ -115,12 +126,17 @@ NS_ASSUME_NONNULL_BEGIN
         // O(n) because it needs to reshift the whole array. So when the breadcrumbs array was full
         // every add operation was O(n).
 
-        _breadcrumbArray[_currentBreadcrumbIndex] = crumb;
+        // Defensive copy: the scope owns an independent snapshot so that (1) callers
+        // mutating the breadcrumb after add don't affect stored data, and (2) when
+        // the ring buffer evicts an old entry its dealloc cannot race with another
+        // thread reading the same object's properties (see #8013).
+        SentryBreadcrumb *snapshot = [crumb snapshotCopy];
+        _breadcrumbArray[_currentBreadcrumbIndex] = snapshot;
 
         _currentBreadcrumbIndex = (_currentBreadcrumbIndex + 1) % _maxBreadcrumbs;
 
         // Serializing is expensive. Only do it once.
-        NSDictionary<NSString *, id> *serializedBreadcrumb = [crumb serialize];
+        NSDictionary<NSString *, id> *serializedBreadcrumb = [snapshot serialize];
         for (id<SentryScopeObserver> observer in [self observerSnapshot]) {
             [observer addSerializedBreadcrumb:serializedBreadcrumb];
         }
@@ -158,6 +174,12 @@ NS_ASSUME_NONNULL_BEGIN
             }
         }
     }
+}
+
+- (void)setPropagationContextWithTraceId:(SentryId *)traceId spanId:(SentrySpanId *)spanId
+{
+    self.propagationContext = [[SentryPropagationContext alloc] initWithTraceId:traceId
+                                                                         spanId:spanId];
 }
 
 - (nullable id<SentrySpan>)span
@@ -210,6 +232,7 @@ NS_ASSUME_NONNULL_BEGIN
     @synchronized(_attributesDictionary) {
         [_attributesDictionary removeAllObjects];
     }
+    [_featureFlagBuffer removeAll];
 
     self.userObject = nil;
     self.distString = nil;
@@ -541,7 +564,11 @@ NS_ASSUME_NONNULL_BEGIN
     traceContext = [self buildTraceContext:span];
     serializedData[@"traceContext"] = traceContext;
 
-    NSDictionary *context = [self context];
+    NSMutableDictionary *context = [self context].mutableCopy;
+    NSDictionary<NSString *, id> *_Nullable featureFlags = [_featureFlagBuffer serializeForContext];
+    if (featureFlags.count > 0) {
+        context[@"flags"] = featureFlags;
+    }
     if (context.count > 0) {
         [serializedData setValue:context forKey:@"context"];
     }
@@ -736,7 +763,7 @@ NS_ASSUME_NONNULL_BEGIN
 {
     if (span != nil) {
         NSDictionary *dict = [SENTRY_UNWRAP_NULLABLE_VALUE(id<SentrySpan>, span) serialize];
-        if (dict[kSentrySpanStatusSerializationKey] != nil) {
+        if (dict[kSentryScopeSpanStatusSerializationKey] != nil) {
             return dict;
         }
 
@@ -747,7 +774,7 @@ NS_ASSUME_NONNULL_BEGIN
         // to any state other than undefined. Spans first have a default status of OK, but we don't
         // want to change this for the trace context status.
         NSMutableDictionary *mutableDict = [dict mutableCopy];
-        mutableDict[kSentrySpanStatusSerializationKey] = kSentrySpanStatusNameOk;
+        mutableDict[kSentryScopeSpanStatusSerializationKey] = kSentrySpanStatusNameOk;
         return mutableDict;
 
     } else {