+++ Tests/SentryTests/MockUIScene.m
@@ -2,6 +2,11 @@
 
 #import "MockUIScene.h"
 #if SENTRY_HAS_UIKIT
+
+@interface MockUISceneSession ()
+@property (nonatomic, copy) UISceneSessionRole sentry_role;
+@end
+
 @implementation MockUIScene
 
 - (instancetype)init
@@ -17,6 +22,25 @@
 
 @end
 
+@implementation MockUISceneSession
+
+- (instancetype)initWithRole:(UISceneSessionRole)role
+{
+    self.sentry_role = role;
+    return self;
+}
+
+- (UISceneSessionRole)role
+{
+    return self.sentry_role;
+}
+
+@end
+
+@interface MockUIWindowScene ()
+@property (nonatomic, strong) MockUISceneSession *sentry_session;
+@end
+
 @implementation MockUIWindowScene
 
 - (instancetype)init
@@ -25,9 +49,20 @@
     return self;
 }
 
+- (instancetype)initWithSessionRole:(UISceneSessionRole)role
+{
+    self.sentry_session = [[MockUISceneSession alloc] initWithRole:role];
+    return self;
+}
+
 - (UISceneActivationState)activationState
 {
     return UISceneActivationStateForegroundActive;
+}
+
+- (UISceneSession *)session
+{
+    return self.sentry_session;
 }
 
 @end